Renamed SupplierSCR module to supplier
