+++ supplier-example/supplier/__init__.py
@@ -3,6 +3,7 @@ The flask application package.
 """
 
 from flask import Flask
+
 app = Flask(__name__)
 
-import SupplierSCR.views
+import supplier.views
