+++ supplier-example/supplier/static/fonts/glyphicons-halflings-regular.woff
