+++ supplier-example/supplier/static/scripts/_references.js
