+++ supplier-example/supplier/static/scripts/jquery-1.10.2.intellisense.js
