+++ supplier-example/supplier/static/scripts/jquery.validate-vsdoc.js
