+++ supplier-example/supplier/static/scripts/jquery.validate.js
