+++ supplier-example/supplier/static/scripts/modernizr-2.6.2.js
