+++ supplier-example/supplier/static/scripts/respond.js
